pyproject.toml
[project]
name = "mogiri"
version = "0.1.0"
description = "Simple local job manager with Web UI"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "apscheduler>=3.10,<4",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
mogiri = "mogiri.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
